+++ CHANGES.md
@@ -2,6 +2,7 @@
 ## Development
 - More efficient create
 - Compatibility with multiple versions of ocaml_intrinsics_kernel
+- Add Builder to construct bitvectors more efficiently
 
 ## 0.1.0.1
 - Simplify append, add test